+++ tests/TelegramBotApiTest.php
@@ -84,6 +84,14 @@ final class TelegramBotApiTest extends TestCase
         $this->assertInstanceOf(User::class, $result);
         $this->assertSame(1, $result->id);
 
+        $decodedResponse = [
+            'ok' => true,
+            'result' => [
+                'id' => 1,
+                'is_bot' => false,
+                'first_name' => 'Sergei',
+            ],
+        ];
         $this->assertSame(
             [
                 [
@@ -91,7 +99,7 @@ final class TelegramBotApiTest extends TestCase
                     'message' => 'Send GET-request "getMe".',
                     'context' => [
                         'type' => 1,
-                        'payload' => '[]',
+                        'payload' => [],
                         'method' => $method,
                     ],
                 ],
@@ -100,17 +108,10 @@ final class TelegramBotApiTest extends TestCase
                     'message' => 'On "getMe" request Telegram Bot API returned successful result.',
                     'context' => [
                         'type' => 2,
-                        'payload' => '{"ok":true,"result":{"id":1,"is_bot":false,"first_name":"Sergei"}}',
+                        'payload' => $decodedResponse,
                         'method' => $method,
                         'response' => $response,
-                        'decodedResponse' => [
-                            'ok' => true,
-                            'result' => [
-                                'id' => 1,
-                                'is_bot' => false,
-                                'first_name' => 'Sergei',
-                            ],
-                        ],
+                        'decodedResponse' => $decodedResponse,
                     ],
                 ],
             ],
@@ -160,7 +161,7 @@ final class TelegramBotApiTest extends TestCase
                     'message' => 'Send GET-request "getMe".',
                     'context' => [
                         'type' => 1,
-                        'payload' => '[]',
+                        'payload' => [],
                         'method' => $method,
                     ],
                 ],
@@ -232,7 +233,7 @@ final class TelegramBotApiTest extends TestCase
                     'message' => 'Send GET-request "getMe".',
                     'context' => [
                         'type' => 1,
-                        'payload' => '[]',
+                        'payload' => [],
                         'method' => $method,
                     ],
                 ],